Add Mapstruct

+++ src/main/java/lu/perso/menuback/services/IngredientServices.java
@@ -2,6 +2,7 @@ package lu.perso.menuback.services;
 
 import lu.perso.menuback.constant.MenuEnum;
 import lu.perso.menuback.data.IngredientEntity;
+import lu.perso.menuback.mappers.IngredientMapper;
 import lu.perso.menuback.models.Ingredient;
 import lu.perso.menuback.repository.DishRepository;
 import lu.perso.menuback.repository.IngredientRepository;
@@ -12,6 +13,7 @@ import org.springframework.util.StringUtils;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 public class IngredientServices {
@@ -21,11 +23,13 @@ public class IngredientServices {
     DishRepository dishRepository;
     @Autowired
     DatabaseServices databaseServices;
+    @Autowired
+    IngredientMapper ingredientMapper;
 
     public List<Ingredient> getAllIngredients() {
         List<IngredientEntity> ingredientList = ingredientRepository.findAll();
         return ingredientList.stream()
-                .map(element -> new Ingredient(element.id(), element.name(), element.sectionId(), element.unit()))
+                .map(element -> ingredientMapper.toView(element))
                 .collect(Collectors.toList());
     }
 
@@ -42,7 +46,7 @@ public class IngredientServices {
                 ingredientModel.unit()
         );
         IngredientEntity savedObject = ingredientRepository.save(newIngredient);
-        return new Ingredient(savedObject.id(), savedObject.name(), savedObject.sectionId(), savedObject.unit());
+        return ingredientMapper.toView(savedObject);
     }
 
     public Ingredient updateIngredient(Ingredient updatedIngredient) throws IllegalStateException {
@@ -66,9 +70,13 @@ public class IngredientServices {
         if (ingredientRepository.findById(ingredientToDelete.id()).isEmpty()) {
             throw new IllegalStateException("This ingredient does not exist");
         }
-        // TODO ERK: Check ingredient does not exist in any recipes
-
-
-        ingredientRepository.deleteById(ingredientToDelete.id());
+        IngredientEntity storedIngredient = ingredientRepository.findById(ingredientToDelete.id())
+                .orElseThrow(() -> new IllegalStateException("This ingredient does not exist"));
+        // Check if ingredient is not used in any dish
+        if (dishRepository.findDishEntitiesByRecipeContaining(Stream.of(storedIngredient).toList()).isEmpty()) {
+            ingredientRepository.deleteById(ingredientToDelete.id());
+        } else {
+            throw new IllegalStateException("This ingredient is used and can not be deleted");
+        }
     }
 }
